[ADD] add battle.js and fix game.js for move feature
