+++ js/map.js
@@ -2,7 +2,7 @@ import keypress from 'keypress';
 
 keypress(process.stdin);
 
-class Point {
+export class Point {
     constructor(x, y) {
         this._x = x;
         this._y = y;
@@ -265,7 +265,7 @@ function BSP( arr, player) {
 }
 
 //화면 출력
-function printBoard(arr) {
+export function printBoard(arr) {
     console.clear();
 
     for (let row of arr) {
@@ -278,89 +278,12 @@ function printBoard(arr) {
     }
 }
 
-// 이동하기
-function move(arr, dx, dy , player) {
-    arr[player.x][player.y] = '·';
 
-    player.x += dx;
-    player.y += dy;
-
-    arr[player.x][player.y] = '●';
-
-    
-    printBoard(arr);
-}
-
-// 타일 탐색
-function checkTile(arr, x, y) {
-    if(arr[x][y] === '%')
-    {
-        console.log(arr[x][y]);
-        return false;
-    }
-    else if(arr[x][y] === '·')
-    {
-        return true;
-    }
-    else{
-        return false;
-    }
-}
-
-// 입력하기
-function userMoveInput(arr, player) {
-
-    return new Promise ((resolve) => {
-        function handleMoveInput(ch, key) {
-            console.log(key.name);
-            if (key) {
-                if (key.name === "up" || ch === '1') { //위로 가기
-                    if(checkTile(arr, player.x -1, player.y + 0)) 
-                        move(arr, -1, 0, player);  
-                    resolve(true);          
-                } else if (key.name === "down") { //아래로 가기
-                    if(checkTile(arr, player.x + 1, player.y + 0)) 
-                        move(arr, 1, 0, player);     
-                    resolve(true); 
-                } else if (key.name === "left") { //왼쪽으로 가기
-                    if(checkTile(arr, player.x + 0, player.y -1)) 
-                        move(arr, 0, -1, player);     
-                    resolve(true); 
-                } else if (key.name === "right") { //오른쪽으로 가기
-                    if(checkTile(arr, player.x + 0, player.y +1)) 
-                        move(arr, 0, 1, player);     
-                    resolve(true); 
-                } else if (key.ctrl && key.name === "c") {
-                    process.exit();
-                }
-            }
-        }
-    
-        // 입력 설정
-        process.stdin.on("keypress", handleMoveInput);
-        process.stdin.setRawMode(true);
-        process.stdin.resume();
-    })
-    
-}
-
-
-async function game() {
-
-    let player = new Point(0,0);
+export function generateMap(player) {
     let arr = Array.from(new Array(MAP_HEIGHT), () => new Array(MAP_WIDTH).fill('%'));
 
     //BSP 알고리즘 실행
     BSP(arr, player);
 
-    await userMoveInput(arr, player);
-
-    console.log("last");
+    return arr;
 }
-
-
-game();
-
-
-
-
